Adding an update_grid function to Qterm, which allows you to update the grid based on x,y,w,h coordinates within the grid system.  The update_grid function will then do the coordinates calculations for either monospace or fontWorkAround scenarios.  Updated calls to QTerm::update to call QTerm::update_grid instead, and modified the parameters accordingly.
Also cleaned up the cursor blinking code so that it will just erase the old cursor, then draw the new cursor in two seperate updates again.  We explicitly turn off the cursor though when we delete the old one first, and turn it back on when we move the cursor.  This at least makes it so the cursor doens't blink while you are moving the cursor keys.  The timer should be restarted though, but not sure how to do that yet.