Better handling of \r
The reason bash was re-displaying the current line and playbook wasn't
was that readline responds to SIGWINCH by sending \r and redisplaying
the line, and terminal wasn't supporting \r.

Fix that by letting \r go back to the beginning of the line, but \n do
both go back to the beginning and go to the next line. I think that
should work properly with both dos and unix newlines.

Conflicts:

	lib/term.c